requirements.txt
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2
numpy>=1.26
Pillow>=10.0
h5py>=3.11
tqdm>=4.66
timm==1.0.13
open-clip-torch==2.30.0
cornet==0.1.0
huggingface-hub>=0.27
